pyproject.toml
[project]
name = "diet"
version = "0.1.0"
description = "DIET (Diet Insight Engine Transformer) - AI-powered service for personalized nutritional analysis, CDC-driven notifications, and dietary recommendations"
authors = [
    { name = "SyntropyHealth", email = "dev@syntropyhealth.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
keywords = ["health", "nutrition", "ai", "medical", "wellness"]

dependencies = [
    "langgraph>=0.2.0",
    "langsmith>=0.1.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.2.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pymilvus>=2.3.0",
    "python-dotenv>=1.0.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "structlog>=23.0.0",
    "rich>=13.0.0",
    "uvicorn>=0.22.0",
    "fastapi>=0.100.0",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.24.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "strict"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
